+++ pocketbase.go
@@ -156,7 +156,7 @@ func (pb *PocketBase) Execute() error {
 
 	done := make(chan bool, 1)
 
-	// wait for interrupt signal to gracefully shutdown the application
+	// listen for interrupt signal to gracefully shutdown the application
 	go func() {
 		sigch := make(chan os.Signal, 1)
 		signal.Notify(sigch, os.Interrupt, syscall.SIGTERM)